Add support for "set autopilot to"
FlightControlManager.cs - Added setter and getter for the bound variable
"autopilot".  This allows the syntax of "set autopilot to [value]" and
"set value to autopilot".  Modified the SelectAutopilotMode method to
support a null or empty string to clear reset the mode back to stability
assist without forcing SAS to on.