Adds an optional restriction `maximumControllableMass`
If set, the CPU is limited to the maximum amount of vessel mass it can
control.  In that situation, it will check the vessel's mass and if the
total mass is too much, it will power down to save power.  Once the
vessel mass drops into the controllable value, it will power the CPU
back up.