READY?! Please everyone give it a go.
- Renamed methods talking about locks so they say UserFunction
- Tested last cases: recursion, calling a library of methods.
- Refactored built-in calls to use $<argstart> now, and return dummy returns.
- Added some test scripts to kerboscript_tests/user_functions

See the test scripts in the PR (kerboscrit_test/user_functions)
to get a feel for what the new feature does
and how you use it.

What I'd really like people to do is regression test it by running
it on their old scripts - try it out on scripts that you knew worked
in the past to ensure they still work now.

One difference I see is that it has enlarged the size of the kRisc
code to such a degree that now KSM files are no longer smaller than
their KS sources.  I'm not sure how much I still care about this
given that we'll be making the limits bigger soon anyway.