Fixed a bug when a script calls another with args.
When one script called another script with args,
the args were getting shifted off by one because
of improper use of $<argstart>.

This was actually a problem I had fixed before, but then I
messed it up again when I did that large refactor for
builtins.

Now it's back to working.