Resolves #732 and Resolves #746. function bodies preparse now.
Function bodies previously didn't get preparsed for triggers
properly.  Now it recurses down into themto ensure it finds
the triggers inside.  This fixes both the problems
because RUN and WHEN are both in need of a preprocess pass to
work properly, and it was the lack of that preprocess pass that
was causing the problem in both these cases.