I suspect that this Resolves #575.
My reason for saying "suspect" is that the conditions
to test to actually see if it's right are so picky
and fiddly that I just don't have time to craft a
cnnvoluted example to try it right now.

If I'm right, then Vessel:ISDEAD will return TRUE in the
rare case where a vessel handle you have refers to a
vessel that *just* died in the last update a moment ago.
(i.e. crashed into the ground after you obtained the
handle to it).