Fixes #925 and changes how lowercasing gets done.
Also required moving how ALT:APOAPSIS and ETA:APOAPSIS
type strings work.

Technically, ALT and ETA are not objects which should mean
they get documented, but as their purpose was to refactor the old
code to keep working... that documentation isn't vital and can
be a separate issue for later.