added suffexes to PartModule.
Quick fix for #947 - this was my quick fix for running script that didn't know what the event names were going to be for a part module. I have compiled and run it and it works. However you will probably want to change the suffix names or maybe implement this a different way.

If you use this I will add docs once the suffix names are decided upon.