Fixes #858 making empty programs have a NOP.
Now all empty program files will get at least 1 instruction
in them - a NOP, which (importantly) is labeled.  Empty programs
did have instructions before, but they didn't come from the
compiler.  The compiler made a zero-length program, and the
opcodes that existed were added afterward by
ProgramBuilder and they just consisted of a push 0, return.
The reason they didn't work is that since they were built outside
the compiler, they had no Label field filled in.  The jump for
a CALL is taken by reading the label of the entrypoint of
the program, which is its zeroth opcode.  When the zeroth
opcode had a missing .Label because it was added ad-hoc after
the compiler was done, the lookup system broke and the result was
returned as zero.

Now the compile simply performs a check.  If the list of opcodes
is length zero after the compile is done, then a dummy NOP is
added by the compiler before it finishes, so there's always at
least one opcode the compiler put there, with all its proper labelling
work that it knows how to do via AddOpcode().