Resolves #40 with new FROM loop syntax.
After much discussion in the slack chat, I settled on this
syntax:

FROM {statements} UNTIL expression STEP {statements} DO {body}

The braces in the FROM and STEP clauses are mandatory, to
help make them stand out, and to avoid the ugly look
of having naked periods in the middle of the statement.  Periods are
necessary because of how statements are defined all over the
place to mandate the period INSIDE the statement rather than
outside it.

(It's too hard now to go back in time and change
that rule to a proper one where the period is defined in a wrapper
rule around statements, and can therefore be left off in cases
like this.  To make that change now requires a major compiler
refactor, as it adds an extra nesting depth level that none
of the VisitBlahNode() methods in the compiler are currently
expecting, and changes their counts of how many nodes to
expect in node.Nodes.Count for certain alternations.)

Therefore since removing the periods is hard at this stage,
leaving the braces optional would have allowed this:

FROM stmt. UNTIL expr STEP stmt. DO {body}

Which looks really messed up with those naked periods in there.

That is why the braces are being made mandatory for the FROM and STEP clauses.

The rule is that the statements in the FROM braces are not REALLY in their own sseparate scope local to those braces (if they were, it would be impossible to use them inside the loop body), but that they will be unrolled and wrapped AROUND the loop body, so the body can see the scope.

The actual scoping turns from this:

FROM {AAAA.} UNTIL BBBB STEP {CCCC.} DO {DDDD.}

into this:

    {
      AAAA.
      UNTIL BBBB {
        DDDD.
        CCCC.
      }
    }

The initializer variables in the FROM are still local to the loop, but it's scope is wrapped *around* the loop so it can be seen inside the loop.