Fixes #289 making triggers interruptable
Now triggers can span multiple FixedUpdates, because
the system memorizes where it was within the ProcessTriggers()
when it hit the IPU limit, and it picks up from there next
time the KOSFixedUpdate happens.

If you create a new trigger (i.e. by one program running another
so it compiled new triggers into the program context) during an
update, the system won't start checking for it to fire until after
it finishes going over the list of triggers that existed the last
time it started on the list.  In other words if you have so many
triggers that they're taking, say, 5 FixedUpdates to finish each
pass through the triggers list, then it will be 5 FixedUpdates
before a newly added trigger starts being checked for.