Fixes #1137 - KSM can now store label-less jumps.
I started to fix it by removing the relative label-less jump
from the compiler, but changed my mind.  If the label-less
jump is a thing you can do in the virtual machine, then KSM files need
to be able to store it correctly.  So instead I worked on changing
it so the KSM file can in fact store the case where a jump or
branch doesn't bother using the label and instead uses just a
relative jump distance only.

Note, if you populate BOTH the label AND the distance, they'd
better agree!! If they don't it's probably wrong in the first place
even before KSM gets involved.  In any case, if you DO populate
both, the KSM file will take the label as official and forget
the relative distance (re-populating it from the label later
when loading, like it does for all the other labeled jumps).