Fixed problem with premature return in function and arg consuming.
I moved the actual consumption of the argbottom mark back into
the OpcodeReturn where it belongs.  Now OpcodeArgBottom merely
peeks at the stack to confirm the argbottom is there, without
consuming it - leaving the consuming of it for the return statement.

This makes it so that if you do a return prematurely, the return
will consume down to the argbottom mark, no matter how far down
the stack that happens to be, but if you don't do a premature
return before the end of the parameters, the OpcodeArgBottom
will still correctly detect parameter mismatches at that point.