Fixes #1151 unlock in trigger.
The fix was small, but also suspicious.  In Compiler.cs, the
method UnlockIdentifier had a check for "if lockObject.IsInitialized()"
that wraps the whole body of the method.  This check is true only
after the lock's initial default dummy function has been built (the
one it will have prior to being assigned in a LOCK statement) by the
compiler.

This created a chicken-and-egg problem.  If the lock is in "normal"
code, but the unlock is in special trigger code, the problem is that
normal code isn't parsed yet when trigger code is being parsed (they're
done in different compiler passes).  So when the UNLOCK was seen by the
compiler, the LOCK it goes with hasn't been parsed yet, and therefore
"if lockObject.IsInitialized()" is false, and so the compiler entirely
skipped the body of UnlockIdentifier() and thus built no Opcodes at all
for the UNLOCK statement.

But if things were parsed in the opposite order - if it parsed mainline
code first and then in a second pass parsed the triggers, then you'd
get the same problem the other way around - then UNLOCKS in main code
couldn't affect LOCKS in triggers.

So thus the chicken-and-egg problem - it needs to parse both kinds of
code before each other to cover all cases.

As I was trying to think of a way around this, I became suspicious of
the need for the "if lockObject.IsInitialized()" check at all.  I
carefully scanned over every single line of code inside the section to
see if any of them would fail if the lock Object wasn't initialized yet.
It looks like the sort of check that's there to protect itself from
executing code that will explode... but none of the methods it was
using seemed to require the lock to have been initialized, as far as
I can tell.  Merely knowing the identifier name was enough, and that
is known the moment the UserFunction() constructor is called, with
or without the initializer code for it having been parsed yet.

As far as I could tell, the check served no purpose whatsoever.

But that seemed suspicious, surely it wouldn't have been inserted
if someone hadn't seen a need for it at some point.

So then I went through a git history log of that section of the code
to determine when that was first added, and why it was there, and
if the body previously contained a dangerous statement that required
it.

Eventually I found that a long time ago, before I revamped locks as
part of making them the same as user functions, the original compiler
code from days of yore from Marianoapp had that statement in it, and
back then it made sense because back then it needed to use a
hash of the compiled initializer code as the ident for the lock.
It no longer needs that, and so that check is no longer desired.

So I removed it.  But be sure to regression test the heck out of it
with your old scripts.
entirely.