Fix Unbinding issues
BindingsUniverse.cs
* Shift to get the instance of STEERINGMANAGER based on shared instead
of shared.Vessel
FlightControlManager.cs
* Switch the call to SteeringManager.RemoveInstance so that it is inside
the Dispose method instead of UnBind (breaking execution causes an
unbind, so ending a script or aborting a script caused a null
reference).
SteeringManager.cs
* Implement a new DeepCopy static method which copies settings and PID
information into a new instance variable.  Mostly for use when a vessel
un-docks and a new SteeringManager needs to be created.
* Fix an error where I was removing elements from a list while
enumerating the list, which throws an exception.