Changed extended ascii Pi char to Latex markup '\pi'
When I edited this file before, it silently changed
the pi character into 'I?' without telling me, because
it wasn't using the same codepage for mapping the
"extended ascii" as whomever had edited it before.

I was going to edit it again with new settings,
but decided it's probably better to just use the sphinx
encoding for pi (which it gets from LateX), '\pi'.