Made RUN parameters stack order same as functions
Previously, it worked like so:

BOTH when calling functions, and running programs, the
caller would push args in stack order like so:

    arg3                 <-- stack top
    arg2
    arg1
    _KOSArgMarkerType_   <-- stack bottom

BUT, between then and when the arguments get read into parameters,
the stack had been massaged by various steps along the way such
that programs and user functions saw the stack differently from
each other.

Program parameters saw a stack like so, and thus had to be read
in reverse:

    arg3                 <-- stack top
    arg2
    arg1
    _KOSArgMarkerType_   <-- stack bottom

while user functions saw a stack like so, and thus had to be
read in normal order:

    arg1                 <-- stack top
    arg2
    arg3
    _KOSArgMarkerType_   <-- stack bottom

This was because RUN ran through two levels of OpcodeCall (which
reverses arguments) while user functions only ran through one
level of it.  (This difference is because of how running a subprogram
runs through the 'maybe compile' function before getting to the
actual program, and interactive RUN had to be made to match that too.)

User functions saw it reversed once, while program arguments saw it
reversed twice so it was back how it started (which is not desired).

In the past this was dealt with by making the DECLARE PARAMETER
statement behave differently depending on if it's inside a function
or inside the main program.  If in the main program, it would read
the arguments in reversed order, while if inside a function it wouldn't.

But in preparation for the changes needed to support optional parameters
with defaults, both kinds of DECLARE PARAMETER (program and user funcs)
have to see the stack the same way, with the lastmost args at the bottom
right before the _KOSArgMarkerType_.  This is because they will use
the _KOSArgMarkerType_ as their flag to tell them they've hit the end
of the args and anything left had better be defaultable or it's an error
that it was missing.

Thus this small refactor changes the design as follows:

Old way: CALLER pushes args the same way regardless of if its RUN
or user func. When it's a run there is one additional reversal that
occurs that doesn't happen when it's a user func call.  So by the time
the stack is seen by the body of the program or user function, the
programs' code is written to expect args in the opposite order of
the order that function's code is written to expect them.

New way: To make the bodies of programs and functions work the same
way as each other in regards to how they expect to see their args,
despite the fact that one contains an additional reversal the other
does not, move the burden of dealing with that onto the caller when
it pushes the args instead.  When the CALLER is setting up args for
a RUN, it will push them in reverse order of how it will when it
is setting up the args for a user function call.  Thus by the time
the bodies of the programs and functions read the args, they are
presented to them uniformly, like so:

    arg1                 <-- stack top
    arg2
    arg3
    _KOSArgMarkerType_   <-- stack bottom

This is in preparation for optional args - to make sure the lastmost
args end up at the bottom of the arg list, next to the arg bottom
marker.