Gimbal calculations using transforms
SteeringManager.cs
* Change the calculation of neutral thrust to get the engine's thrust
vector based on the changed gimbal rotation, instead of trying to
reverse engineer the rotation difference.
* Thrust vectors calculate torque based on the gimbal position.
* IMPORTANT: This means that if an engine has a thrust transform that is
not aligned with the gimbal position, torque may be inaccurate.  This
change was forced by the "Vector" engine which is the only stock engine
which changes it's thrust transform location drastically with the
gimbal.
* Attempt to "future proof" handling of gimbals in case an engine ends
up with one gimbal transform controlling more than one thrust transform.
* Change the torque calculation to use the force direction instead of
the engine exhaust direction.  Because we are using the absolute value
of the torque, it didn't change performance but it is more technically
correct.