Now prog params same as func params in compiler.cs
Now the Compiler.cs code path is as close to being
the same as is possible for both prog parameters and
function parameters.  They both route through the same
code with the one exception being that programs' bodies
are dealt with in VisitStartStatement() and functions'
bodies are dealt with in VisitInstructionBlock().  That
part still has to be separate because the list of
instructions that make up a program block is not parsed
with the same grammar rule as the list of instructions
that make up a function block, and the compiler has to
match the grammar rules one-to-one.