Made it possible to call delegates directly().
The ":call" suffix is now not required.  You can
use parens directly if you like.

The only difference now between delegates and native
functions is the syntax and what happens when you
leave the parens off.

When you do this:

function y { return 5. }
set x to y@.

Then x() and y() are the same thing, and x@ and y@ are the same,

but they differ what happens when you leave them bare.

y, bare, means the same as y()
x, bare, means the same as x@.

This difference is very hard to remove, unfortunately.