Fixes #1328 Escape in telnet.
The problem was caused by the code presuming every time there
was an ESC that there MUST be another character following it
that it can read ahead on the next index.  When there wasn't
yet and the snippet of input chars just ended at the ESC, the
attempt to go one more character forward threw an array bounds
exception.  This was never sent to the output_log, however
because the telnet server runs in a separate thread and
apparently the output_log only contains the main thread's output.

When the telnet server thread died from the exception, the telnet
client doesn't show anything about it, other than just freezing
for several seconds and then eventually reporting a dropped
connection.