KSP 1.1 build 1209 required small change to compile.
The change was that KSP build 1209 changed the interfaces:

    IPartMassModifier
    IPartCostModifier

So they both needed one new additional method, and one
new additional parameter to the method they already had.
Since this was just related to parts that shed part of themselves
away when staging and keep part of themselves behind (fairings),
for our purpopses the changes were irrelevant and we could just
ignore the extra information.  But we had to add dummies for them
anyway just to comply with the new interface definition so it
would compile.