Convert value on "Set" rather than during autopilot
SteeringManager.cs
* Move conversion of the "set" value from inside the autopilot update
and into the actual set method itself.  This means that a conversion
exception will break execution and display the error (on the next
physics tick) preventing further script execution.

KSPLogger.cs
* Remove requirement for null or empty filepath for noting that an
exception is in the built-in steering update (previously this line was
not getting triggered unless called from the interpreter).