Check volume range in ProcessElectricity
Fixes #1363

kOSProcessor.cs
* Add a check to the volume range before accessing the current volume in
Process electricity.
* This allows "start on archive" to properly work with RemoteTech, and
will simply skip the boot sequence if out of range.
* Add Archive property for storing the archive volume after creation.
* Log exception when boot volume is out of range (so that it is
displayed on the terminal too).
* Change CheckCanBoot logger calls to be shared.Logger, so that the
message displays on the terminal.

KOSVolumeOutOfRangeException
* Create new exception class to handle "volume out of range" exceptions.

RemoteTechVolumeManager.cs
* Use the new exception type.

kOS.Safe.csproj
* Add reference for new exception.

KSPLogger.cs
* Add null checks before attempting to create the code fragment message.
If called from the SafeHouse instance, Shared may be null.
* Remove call to base.Log(text) in Log(string text).  This forced the
logger to attempt to display the text on the screen as well as in the
log.  But the screen printing feature never got used, because the method
was only called on the SafeHouse instance (which has no Shared object
and therefore no Screen).

Logger.cs
* Add null checking to audible exceptions, because a call to the
SafeHouse instance may have a null Shared object.