add orbit constructor and velocityat/positionat
adds a constructor for Orbit (OrbitInfo internally) structures so
that users can construct their own orbit objects.

this lets a function that can build a hohmann transfer node to the
mun's orbit also be reused to transfer to a keosynchronous orbit.

because the orbit structure has no associated orbitable body it is
not possible to use positionat(orbitable) and velocityat(orbitable)
functions, so therefore added positionat and velocityat functions
to the orbit object (APIs that do not take into account maneuver
nodes and are documented as such).

docs updates and kerboscript tests included.