Fixed bug due to diff between mono and C# compiler
Microsoft's C# compiler handled anonymous delegates inside
a static method a bit differently than Mono's did, meaning
this code that worked properly when compiled in Visual Studio
didn't work when compiled in Mono compilers.  This change
unrolls some of C#'s under the hood "optomizations" and
turned them into explicit code.

I say "optomizations" in quotes because it's a lot more
than just an optomization if it actually makes the end
result different in some way other than just its speed or
memory footprint, as was the case here.  The thing the C#
compiler was doing to the code to slightly alter its
meaning was actually necessary to the functioning of the
call to the KSP API.  (It doesn't work when given a static
delegate as the source code looked like it was doing, but
the C# optomizations were changing it so it wasn't quite the
static delegate it appeared to be.) Since this was necessary to
keep the code from crashing, that "optomization" should be made
explicit in our code and not be dependent on differing compiler
behaviours.

With this edit, the code now works the same way on Microsoft's
or on Mono's compilers.