Distinguishes line 13, col 2, from line 1, col 32
I just realized a potential failure case in the very
odd specific instance where the same function
body exists in two places that when line and col
get concatenated happen to form the same string.

i.e. line 13, column 2, when concatenated make the string 132.
But so would line 1, column 32.  It would also be 132.

So I added a few chars to the format string to distinguish them.
Now line 13, column 2 will be L13C2 and line 1, column 32 will
be L1C32.