ITERATOR:RESET not supported anymore.
I found out that we had to remove the :RESET suffix entirely
because the way the new collections structures we have
no longer actually support rewinding the enumerators:

When we went with the newer collections style, there
was a revamp to go through all of them and change how
they returned ITERATORS.  Instead of returning a hand-made
IEnumerator, the technique was changed to the one where you
let C# magically create an IEnumerator<fooType> out of
a single method where all it does a yield return <fooType>.

Apparently this technique causes it to build an IEnumerator
in which the .Reset() method has been defined as just

public void Reset()
{
   throw new System.NotSupportedException().
}

So this part of our docs have been lying ever since we made that
change.  The RESET was never implemented an never worked ever
since then.