Review edits
VoiceValue.cs
* clean up a couple of white space issues
* spelling fixes in one of the comments

Merge branch 'develop' into pull-review/1915

Fixes #1913 by timing notes when they should go.

The fix is to time the
next note based on when the previous
note *should* have ended, not based
on when it *did*.
Notes get cut shrot if they started late, to get back
on
track with the 'sheet music'.  If the lag gets really bad
it may end
up skipping an entire note.

To test this, try the example script in
issue #1913 and
notice how it's different before and after this PR is
merged.

After this PR is merged, the song reliably gives 3 bongs
between
Bings even when the physics rate is slowed down.
(before the
lagging caused the count to be off so it sometimes
gave 1 or 2 bongs
instead of 3 between Bings.)

Added a debugfreezegame built-in function,

for use when trying to force
laggy situations when testing.
I did not make this specific to having to
compile in
DEBUG mode. It's there all the time.  I figure it might
be
useful for the users to have access to as well.

(However, it's not
documented, on purpose for now, in the
Sphinx docs, until we get a
consensus on whether or not
to make it public.)

Included here is an
example script that uses it to
force an instance of issue #1913

Merge branch 'refs/heads/pull-review/1915' into develop

Review edits

VoiceValue.cs
* clean up a couple of white space issues
*
spelling fixes in one of the comments

Merge branch 'develop' into pull-review/1915