Added some new suffixes for gui callbacks.
Changes:

1 - Added Popup hook OnChange.  Only fires when the value differs
from whatever it was before.  (If you pick the same value, there
is no callback fired).  Also fiddled with making Popup's into
combolistboxes that can scroll, by adding the concept of the
MAXVISIBLE suffix that will add scrollbars when the number of
items in the list is bigger than MAXVISIBLE.  This involved
some pixel size calculations I'm not 100% sure about.

2 - Renamed button:onchange to button:ontoggle because of the
fact that Popup lists are also a kind of Button so it got
confusing to have two different meanings of ONCHANGE for Popups
(where there's the button meaning (when the button goes in or
goes out) and the list selection meaning (when the value printed
on the face of the button changes because of the list pick)).
Now the meanings are separated.  OnTOGGLE is the button state,
and OnChange only exists with Popups, not normal buttons.  If
you have a Popup, then it can have both Onchange (for the list
value pick) and onToggle( for the button state change) hooks.

3 - Added ONCHANGE hook support to vertical and horizontal
sliders, and added a current value field to the constructors
for such.