Adds "uniques heap" for VesselTarget instances.
Fixes #1980 and #1982.

Problem #1982 was simpler so I'll explain it first:

When the SteeringManager complains about the CPU controlling
an already-controlled craft, it throws an exception which
makes the CPU try to abort execution.  But as part of trying to
abort that execution it calls PopContext(), which in turn
calls the SteeringManager to ask it to unbind the steering,
where the SteeringManager again throws an exception because
of the dual CPU control.  This second exception makes the CPU
abort partway through finishing the work of breaking the
execution, so it never quite fully terminates the program, which
is why it keeps trying to run after some of the running context
information has been wiped away (thus the null refs in
ExecuteInstruction).

The fix:  The CPU now advertises when it's in a PopContext,
so the SteeringManager can choose not to throw that second
exception in that case.

Problem #1980 required a slight refactor:

This problem was caused by the fact that VesselTarget now
inserts callback hooks into GameEvents, which it didn't do
before, and those callback hooks prevent it from getting
garbage collected.   As long as GameEvents has a
reference to a delegate that needs the VesselTarget object
to keep existing as part of its closure, the VesselTarget
can't ever get orphaned.  It was depending on being a
KOSScopeObserver to discover when kOS orphans it, so it
can remove its hooks from GameEvents, but KOSScopeObserver
can't really reliably catch all cases like that, and relies
on a C# destructor ( ~Variable ) to call ScopeLost().  But
the C# destructor never happens if the thing never orphans,
so it's a catch-22.  The orphaning event that tells
VesselTarget to unhook itself won't happen because VesselTarget
hasn't unhooked itself.

Thus the freeze at docking was caused by the fact that
every time the docking script had referred to the target vessel
during its main loop, it was creating a new instance of the
VesselTarget wrapper around it, and the old instances
never truly orphaned away like they should.  By the time
the docking happened there were thousands of thse extra
instances in memory, and upon docking ALL their callback
hooks got fired off at once.

The Fix:

Since VesselTarget is just a wrapper around a KSP Vessel,
and all the information contained therein is meant to
just inherit the info from the Vessel, and all the
Setter Suffixes are meant to pass the change on to the
KSP Vessel itself, there is no practical difference between
a kerboscript having two different VesselTarget instances
wrapping the same Vessel versus having just one VesselTarget
instance for it and then just having two references to it.

Threfore we track whether or not a VesselTarget wrapper
around a Vessel already exists, and if it does then we just
hand out a reference to the existing one instead of constructing
a new one.  That prevents us from having thousands of copies
of the same VesselTarget in memory when they don't orphan.

It should also have the nice side effect of reducing
execution burden by not re-running the Parts query of
the vessel each time it's referrenced in the script, as
it had to do before when each mention of TARGET made a
brand new VesselTarget.

The reason so many files got touched is because all instances
of calling the VesselTarget constructor had to be edited
to use a factory method instead, that does what the
above paragraphs describe.