Boolean meaning 0.00001 = true, 0.00000 = false
I moved the code out of abstract class ScalarValue
into the classes that override it, because I wanted
to use a simple cast operator and to do that I need
to know the type of Value at compile-time.

I could have just used C#'s own knowledge of how you're
meant to "convert" numbers to Booleans by doing
Convert.ToBoolean(Value) in ScalarValue.  It would have
given the same result, but I think this way is a bit faster
because it casts instead of using Convert methods.